+++ pyannote/metrics/diarization.py
@@ -51,6 +51,18 @@ class DiarizationErrorRate(IdentificationErrorRate):
     error rate is computed as the identification error rate with each hypothesis
     label translated into the corresponding reference label.
 
+    Parameters
+    ----------
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
+
+    Usage
+    -----
+
     * Diarization error rate between `reference` and `hypothesis` annotations
 
         >>> metric = DiarizationErrorRate()
@@ -86,8 +98,9 @@ class DiarizationErrorRate(IdentificationErrorRate):
     def metric_name(cls):
         return DER_NAME
 
-    def __init__(self, **kwargs):
-        super(DiarizationErrorRate, self).__init__(**kwargs)
+    def __init__(self, collar=0.0, skip_overlap=False, **kwargs):
+        super(DiarizationErrorRate, self).__init__(
+            collar=collar, skip_overlap=skip_overlap, **kwargs)
         self.mapper_ = HungarianMapper()
 
     def optimal_mapping(self, reference, hypothesis, uem=None):
@@ -121,8 +134,10 @@ class DiarizationErrorRate(IdentificationErrorRate):
 
         # crop reference and hypothesis to evaluated regions (uem)
         # remove collars around reference segment boundaries
+        # remove overlap regions (if requested)
         reference, hypothesis, uem = self.uemify(
-            reference, hypothesis, uem=uem, collar=self.collar,
+            reference, hypothesis, uem=uem,
+            collar=self.collar, skip_overlap=self.skip_overlap,
             returns_uem=True)
         # NOTE that this 'uemification' must be done here because it
         # might have an impact on the search for the optimal mapping.
@@ -138,10 +153,11 @@ class DiarizationErrorRate(IdentificationErrorRate):
 
         # compute identification error rate based on mapped hypothesis
         # NOTE that collar is set to 0.0 because 'uemify' has already
-        # been applied
+        # been applied (same reason for setting skip_overlap to False)
         mapped = hypothesis.translate(mapping)
         return super(DiarizationErrorRate, self)\
-            .compute_components(reference, mapped, uem=uem, collar=0.0,
+            .compute_components(reference, mapped, uem=uem,
+                                collar=0.0, skip_overlap=False,
                                 **kwargs)
 
 
@@ -152,6 +168,18 @@ class GreedyDiarizationErrorRate(IdentificationErrorRate):
     obtained. Then, the actual diarization error rate is computed as the
     identification error rate with each hypothesis label translated into the
     corresponding reference label.
+
+    Parameters
+    ----------
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
+
+    Usage
+    -----
 
     * Greedy diarization error rate between `reference` and `hypothesis` annotations
 
@@ -187,8 +215,9 @@ class GreedyDiarizationErrorRate(IdentificationErrorRate):
     def metric_name(cls):
         return DER_NAME
 
-    def __init__(self, **kwargs):
-        super(GreedyDiarizationErrorRate, self).__init__(**kwargs)
+    def __init__(self, collar=0.0, skip_overlap=False, **kwargs):
+        super(GreedyDiarizationErrorRate, self).__init__(
+            collar=collar, skip_overlap=skip_overlap, **kwargs)
         self.mapper_ = GreedyMapper()
 
     def greedy_mapping(self, reference, hypothesis, uem=None):
@@ -215,11 +244,13 @@ class GreedyDiarizationErrorRate(IdentificationErrorRate):
 
         # crop reference and hypothesis to evaluated regions (uem)
         # remove collars around reference segment boundaries
+        # remove overlap regions (if requested)
         reference, hypothesis, uem = self.uemify(
-            reference, hypothesis, uem=uem, collar=self.collar,
+            reference, hypothesis, uem=uem,
+            collar=self.collar, skip_overlap=self.skip_overlap,
             returns_uem=True)
         # NOTE that this 'uemification' must be done here because it
-        # might have an impact on the search for the optimal mapping.
+        # might have an impact on the search for the greedy mapping.
 
         # make sure reference only contains string labels ('A', 'B', ...)
         reference = reference.anonymize_labels(generator='string')
@@ -232,11 +263,12 @@ class GreedyDiarizationErrorRate(IdentificationErrorRate):
 
         # compute identification error rate based on mapped hypothesis
         # NOTE that collar is set to 0.0 because 'uemify' has already
-        # been applied
+        # been applied (same reason for setting skip_overlap to False)
         mapped = hypothesis.translate(mapping)
         return super(GreedyDiarizationErrorRate, self)\
-            .compute_components(reference, mapped, uem=uem, collar=0.0, **kwargs)
-
+            .compute_components(reference, mapped, uem=uem,
+                                collar=0.0, skip_overlap=False,
+                                **kwargs)
 
 PURITY_NAME = 'purity'
 PURITY_TOTAL = 'total'
@@ -253,6 +285,12 @@ class DiarizationPurity(UEMSupportMixin, BaseMetric):
     ----------
     weighted : bool, optional
         When True (default), each cluster is weighted by its overall duration.
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
     """
 
     @classmethod
@@ -263,19 +301,24 @@ class DiarizationPurity(UEMSupportMixin, BaseMetric):
     def metric_components(cls):
         return [PURITY_TOTAL, PURITY_CORRECT]
 
-    def __init__(self, weighted=True, **kwargs):
+    def __init__(self, collar=0.0, skip_overlap=False,
+                 weighted=True, **kwargs):
         super(DiarizationPurity, self).__init__(**kwargs)
         self.weighted = weighted
+        self.collar = collar
+        self.skip_overlap = skip_overlap
 
     def compute_components(self, reference, hypothesis, uem=None, **kwargs):
 
         detail = self.init_components()
 
         # crop reference and hypothesis to evaluated regions (uem)
-        reference, hypothesis = self.uemify(reference, hypothesis, uem=uem)
+        reference, hypothesis = self.uemify(
+            reference, hypothesis, uem=uem,
+            collar=self.collar, skip_overlap=self.skip_overlap)
 
         # cooccurrence matrix
-        matrix = reference.support() * hypothesis.support()
+        matrix = reference * hypothesis
 
         # duration of largest class in each cluster
         largest = matrix.max(dim='i')
@@ -312,14 +355,23 @@ class DiarizationCoverage(DiarizationPurity):
     ----------
     weighted : bool, optional
         When True (default), each cluster is weighted by its overall duration.
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
     """
 
     @classmethod
     def metric_name(cls):
         return COVERAGE_NAME
 
-    def __init__(self, weighted=True, **kwargs):
-        super(DiarizationCoverage, self).__init__(weighted=weighted, **kwargs)
+    def __init__(self, collar=0.0, skip_overlap=False,
+                 weighted=True, **kwargs):
+        super(DiarizationCoverage, self).__init__(
+            collar=collar, skip_overlap=skip_overlap,
+            weighted=weighted, **kwargs)
 
     def compute_components(self, reference, hypothesis, uem=None, **kwargs):
         return super(DiarizationCoverage, self)\
@@ -339,14 +391,18 @@ class DiarizationPurityCoverageFMeasure(UEMSupportMixin, BaseMetric):
     Parameters
     ----------
     weighted : bool, optional
-        When True (default), each cluster/class is weighted by its overall duration.
+        When True (default), each cluster/class is weighted by its overall
+        duration.
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
     beta : float, optional
         When beta > 1, greater importance is given to coverage.
         When beta < 1, greater importance is given to purity.
         Defaults to 1.
-
-    Note
-    ----
 
     See also
     --------
@@ -367,8 +423,11 @@ class DiarizationPurityCoverageFMeasure(UEMSupportMixin, BaseMetric):
                 PURITY_COVERAGE_LARGEST_CLUSTER,
                 PURITY_COVERAGE_TOTAL_CLASS]
 
-    def __init__(self, weighted=True, beta=1., **kwargs):
+    def __init__(self, collar=0.0, skip_overlap=False,
+                 weighted=True, beta=1., **kwargs):
         super(DiarizationPurityCoverageFMeasure, self).__init__(**kwargs)
+        self.collar = collar
+        self.skip_overlap = skip_overlap
         self.weighted = weighted
         self.beta = beta
 
@@ -377,10 +436,12 @@ class DiarizationPurityCoverageFMeasure(UEMSupportMixin, BaseMetric):
         detail = self.init_components()
 
         # crop reference and hypothesis to evaluated regions (uem)
-        reference, hypothesis = self.uemify(reference, hypothesis, uem=uem)
+        reference, hypothesis = self.uemify(
+            reference, hypothesis, uem=uem,
+            collar=self.collar, skip_overlap=self.skip_overlap)
 
         # cooccurrence matrix
-        matrix = reference.support() * hypothesis.support()
+        matrix = reference * hypothesis
 
         # duration of largest class in each cluster
         largest_class = matrix.max(dim='i')
@@ -448,7 +509,18 @@ HOMOGENEITY_CROSS_ENTROPY = 'cross-entropy'
 
 
 class DiarizationHomogeneity(UEMSupportMixin, BaseMetric):
-    """Cluster homogeneity"""
+    """Cluster homogeneity
+
+    Parameters
+    ----------
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
+
+    """
 
     @classmethod
     def metric_name(cls):
@@ -458,27 +530,36 @@ class DiarizationHomogeneity(UEMSupportMixin, BaseMetric):
     def metric_components(cls):
         return [HOMOGENEITY_ENTROPY, HOMOGENEITY_CROSS_ENTROPY]
 
+    def __init__(self, collar=0.0, skip_overlap=False, **kwargs):
+        super(DiarizationHomogeneity, self).__init__(**kwargs)
+        self.collar = collar
+        self.skip_overlap = skip_overlap
+
     def compute_components(self, reference, hypothesis, uem=None, **kwargs):
 
         detail = self.init_components()
 
         # crop reference and hypothesis to evaluated regions (uem)
-        reference, hypothesis = self.uemify(reference, hypothesis, uem=uem)
+        reference, hypothesis = self.uemify(
+            reference, hypothesis, uem=uem,
+            collar=self.collar, skip_overlap=self.skip_overlap)
 
         # cooccurrence matrix
-        matrix = reference.support() * hypothesis.support()
+        matrix = np.array(reference * hypothesis)
 
-        duration = matrix.sum()
-        rduration = matrix.sum(dim='j')
-        hduration = matrix.sum(dim='i')
+        duration = np.sum(matrix)
+        rduration = np.sum(matrix, axis=1)
+        hduration = np.sum(matrix, axis=0)
 
         # reference entropy and reference/hypothesis cross-entropy
-        ratio = rduration / duration
-        entropy = -(ratio * np.log(ratio)).sum()
-        detail[HOMOGENEITY_ENTROPY] = entropy.item()
+        ratio = np.ma.divide(rduration, duration).filled(0.)
+        detail[HOMOGENEITY_ENTROPY] = \
+            -np.sum(ratio * np.ma.log(ratio).filled(0.))
 
-        cross_entropy = -((matrix / duration) * np.log(matrix / hduration)).sum()
-        detail[HOMOGENEITY_CROSS_ENTROPY] = cross_entropy.item()
+        ratio = np.ma.divide(matrix, duration).filled(0.)
+        hratio = np.ma.divide(matrix, hduration).filled(0.)
+        detail[HOMOGENEITY_CROSS_ENTROPY] = \
+            -np.sum(ratio * np.ma.log(hratio).filled(0.))
 
         return detail
 
@@ -498,7 +579,18 @@ COMPLETENESS_NAME = 'completeness'
 
 
 class DiarizationCompleteness(DiarizationHomogeneity):
-    """Cluster completeness"""
+    """Cluster completeness
+
+    Parameters
+    ----------
+    collar : float, optional
+        Duration (in seconds) of collars removed from evaluation around
+        boundaries of reference segments.
+    skip_overlap : bool, optional
+        Set to True to not evaluate overlap regions.
+        Defaults to False (i.e. keep overlap regions).
+
+    """
 
     @classmethod
     def metric_name(cls):